functionize `getFilesContent()`
