+++ src/components/ChatCompletion.jsx
@@ -4,7 +4,7 @@ import { useMsal } from '@azure/msal-react';
 import { AzureOpenAI } from 'openai';
 
 import { loginRequest } from '../authConfig';
-import { getEmail, getFileList, getChatMessages, getChannelMessageList } from '../graph';
+import { getEmail, getFilesContent, getChatMessages, getChannelMessageList } from '../graph';
 import { ChatCompletionData } from '../dataview';
 import { systemPrompt, userPrompt } from '../values';
 
@@ -58,30 +58,10 @@ export const ChatCompletion = () => {
 
                     const email = getEmail(token);
 
-                    const file_content = getFileList(token, file_path)
-                        .then((file_list) => {
-                            const urls = file_list.map(d => d["@microsoft.graph.downloadUrl"]);
-
-                            async function get_content(url, callback) {
-                               const config = {
-                                    newlineDelimiter: " ",
-                                    ignoreNotes: true
-                                }
-                                const response = await fetch(url);
-                                const arrayBuffer = await response.arrayBuffer();
-                                const result = await officeParser.parseOfficeAsync(arrayBuffer, config);
-                                return(result);
-                            }
-
-                            return Promise.all(urls.map(a => get_content(a)))
-                                .then((text) => {
-                                    const result = file_list.map((e,i) => ({
-                                        ...e,
-                                        text: text[i]
-                                    }));
-                                    return(result);
-                                })
-                        });
+                    let file_content = Promise.resolve([]);
+                    if (file_path !== null && file_path != "") {
+                        file_content = getFilesContent(token, file_path);
+                    }
 
                     let chat_msgs = Promise.resolve({value:[]});
                     if (selected_chats !== null) {
