+++ src/graph.js
@@ -55,13 +55,47 @@ export async function getChatMessages(accessToken, chat_id, daysBefore = daysBef
 }
 
 export async function getFileList(accessToken, file_path, daysBefore = daysBefore_global) {
-    const url = "https://graph.microsoft.com/v1.0/me/drive/root:/" + file_path + ":/children";
-    const dir_list = await getGraphResponse(accessToken, url);
+    const api_url =
+        "https://graph.microsoft.com/v1.0/me/drive/root:/" +
+        file_path +
+        ":/children";
+    const dir_list = await getGraphResponse(accessToken, api_url);
+
     const recent_dir_list = dir_list.value
         .filter(e => e.lastModifiedDateTime > getStartFromDateStr(daysBefore));
-    const file_list = recent_dir_list.filter(e => !e.folder);
+    const file_list = recent_dir_list.filter(e => e.file);
     const subfolder_list = recent_dir_list.filter(e => e.folder);
+
     return file_list;
+}
+
+export async function getFileContent(accessToken, file_url) {
+    const config = {
+        newlineDelimiter: " ",
+        ignoreNotes: true
+    };
+    const response = await fetch(file_url);
+    const arrayBuffer = await response.arrayBuffer();
+    const result = await officeParser.parseOfficeAsync(arrayBuffer, config);
+    return result;
+}
+
+export async function getFilesContent(accessToken, file_path, daysBefore = daysBefore_global) {
+    const file_list = await getFileList(accessToken, file_path, daysBefore);
+
+    // filter to files that officeParser can parse
+
+    const urls = file_list.map(d => d["@microsoft.graph.downloadUrl"]);
+
+    const result = await Promise.all(urls.map(a => getFileContent(accessToken, a)))
+        .then((text) => {
+            return file_list.map((e, i) => ({
+                ...e,
+                text: text[i]
+            }));
+        });
+
+    return result;
 }
 
 export async function getTeamList(accessToken) {
